deprecate Pyblosxom support
Since no one is going to main the Pyblosxom part of code, it's best for
retirement for that. Futuremore, smartypants.py should keep generic rather
than binding with specific softwares.

It should be very easy to wrap smartypants.py and make a new plugin.

The support will be removed at Version 2.0.0