use re.match instead of re.search to improve performance
if there isn't a match to the second half of the tag regex, the regex
engine backtracks through the entire string character by character.
one way to solve this is to use match instead of search. search, if it
doesn't find a match starting with char 0, tries again starting at char
1, and so on and so forth - so if there's no match it'll take O(n^2)
polynomial time to complete. match, on the other hand, only matches if
the entire string matches. Since we start off with a greedy capturing
group to consume everything before a tag, this is safe to use.