Fix double closing quotes after an HTML block
The double closing quote happens to be at the beginning. If it is
followed by a space, everything is OK, but if it is followed by a
punctuation, we get the default opening quote instead. Fix that by
using closing tag when there is punctuation.

This is not really generic because we do not want to break other
cases, like quoting a punctuation character.